+++ packages/onchain-role-enrolment/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
- All notable changes to this project will be documented in this file.
 
- The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to
+# 1.0.0 (10-02-2022)
+
+### Features
+
+* Code migration from ([iam-contracts](https://github.com/energywebfoundation/iam-contracts))
+* This package contains ClaimManager and On-chain RevocationRegistry contract and it's client code.